02-security/opcua-app/requirements.txt
asyncua>=1.0.0
cryptography
RPi.GPIO
adafruit-circuitpython-bme680